+++ lib/df_searchable_dropdown.dart
@@ -1,7 +1,7 @@
-import 'package:df_dropdown/enums/dropdown_type.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
+import '/enums/dropdown_type.dart';
 import '/models/drop_down_model.dart';
 import '/models/dropdown_decoration.dart';
 import '/models/simple_selector_decoration.dart';
@@ -60,9 +60,13 @@ class DfSearchableDropdown<T> extends StatelessWidget {
   final String? hintText;
 
   /// Callback triggered when an option from the dropdown is selected.
-  final Function(DropDownModel<T>)? onOptionSelected;
+  final Function(DropDownModel<T>?)? onOptionSelected;
 
-  /// Validator function for validating the selected dropdown option.
+  /// Provides a [DropDownModel] object if selected, and `null` if not
+  ///
+  /// Should return `null` when no validation error is present,
+  /// and a [String] if there is an error
+  ///
   final String? Function(DropDownModel<T>?)? validator;
 
   /// Function that performs the search operation based on the user's input. It returns a list of filtered options.
@@ -157,13 +161,24 @@ class _DropdownState<T> extends State<_Dropdown<T>> {
   Widget build(BuildContext context) {
     final provider =
         Provider.of<SearchableDropdownProvider<T>>(context, listen: false);
-    final textTheme = Theme.of(context).textTheme;
 
     return Column(
       children: [
         DropdownField<SearchableDropdownProvider<T>>(
           disabled: widget.disabled,
-          key: provider.dropdownKey,
+          onEditingComplete: () {
+            if (provider.searchTextController.text.isEmpty) {
+              provider.onSelectSuggestion(null);
+            } else if (provider.selectedValue != null &&
+                provider.selectedValue?.text !=
+                    provider.searchTextController.text) {
+              provider.searchTextController.text = provider.selectedValue!.text;
+            }
+
+            FocusScope.of(context).requestFocus(FocusNode());
+            provider.closeSuggestions();
+          },
+          dropdownType: widget.dropdownType,
           decoration: widget.decoration,
           hintText: widget.hintText,
           labelText: widget.labelText,
@@ -182,14 +197,16 @@ class _DropdownState<T> extends State<_Dropdown<T>> {
           },
           suffixTapEnabled: false,
           suffixWidget: GestureDetector(
-            onTap: () => provider.toggleSuggestionsExpanded(
-              selectorWidget: widget.dropdownType == DropdownType.expandable
-                  ? null
-                  : ChangeNotifierProvider.value(
-                      value: provider,
-                      child: selectorWidget,
-                    ),
-            ),
+            onTap: () {
+              provider.toggleSuggestionsExpanded(
+                selectorWidget: widget.dropdownType == DropdownType.expandable
+                    ? null
+                    : ChangeNotifierProvider.value(
+                        value: provider,
+                        child: selectorWidget,
+                      ),
+              );
+            },
             child: SizedBox(
               height: 48,
               child: widget.arrowWidget ??
@@ -209,23 +226,6 @@ class _DropdownState<T> extends State<_Dropdown<T>> {
           ),
           selectorWidget,
         ],
-        //An error message [Text] widget displayed only when validation returns an error
-        if ((!provider.suggestionsExpanded &&
-                widget.dropdownType == DropdownType.expandable) ||
-            (widget.decoration?.reserveSpaceForValidationMessage != false ||
-                provider.validationError != null))
-          Padding(
-            padding: const EdgeInsets.only(
-              top: 4,
-            ),
-            child: Text(
-              provider.validationError ?? "",
-              style: widget.decoration?.errorMessageTextStyle ??
-                  textTheme.bodySmall?.copyWith(
-                    color: Colors.red.shade500,
-                  ),
-            ),
-          )
       ],
     );
   }