+++ lib/providers/searchable_dropdown_provider.dart
@@ -19,10 +19,13 @@ class SearchableDropdownProvider<T> extends BaseDropdownProvider<T> {
   }
 
   DropDownModel<T>? selectedValue;
-  final List<DropDownModel<T>> searchResults = [];
-  final Function(DropDownModel<T>)? onOptionSelected;
+  final List<DropDownModel<T>> _searchResults = [];
+  final Function(DropDownModel<T>?)? onOptionSelected;
   final Future<List<DropDownModel<T>>> Function(String searchText)? onSearch;
   final double? selectorMaxHeight;
+
+  List<DropDownModel<T>> get searchResults =>
+      [if (selectedValue != null) selectedValue!, ..._searchResults];
 
   @override
   double get dropdownHeight {
@@ -39,9 +42,9 @@ class SearchableDropdownProvider<T> extends BaseDropdownProvider<T> {
     double height = 0;
 
     int dataLength =
-        searchResults.isNotEmpty || searchTextController.text.isNotEmpty
+        (searchResults.isNotEmpty || searchTextController.text.isNotEmpty
             ? searchResults.length
-            : initData.length;
+            : initData.length);
 
     if (dataLength < 5) {
       height = dataLength * 40;
@@ -49,14 +52,17 @@ class SearchableDropdownProvider<T> extends BaseDropdownProvider<T> {
       height = selectorMaxHeight ?? 200;
     }
 
-    return height;
+    return height > 0 ? height : 40;
   }
 
-  void onSelectSuggestion(DropDownModel<T> value) {
+  void onSelectSuggestion(DropDownModel<T>? value) {
+    if (value == selectedValue) return;
     selectedValue = value;
     validationError = null;
     closeSuggestions();
-    searchTextController.text = value.text;
+    if (value != null) {
+      searchTextController.text = value.text;
+    }
     if (onOptionSelected != null) {
       onOptionSelected!(value);
     }
@@ -73,17 +79,26 @@ class SearchableDropdownProvider<T> extends BaseDropdownProvider<T> {
   }
 
   @override
+  void closeSuggestions() {
+    if (selectedValue != null &&
+        selectedValue?.text != searchTextController.text) {
+      searchTextController.text = selectedValue!.text;
+    }
+    super.closeSuggestions();
+  }
+
+  @override
   void onInputChanged(String text) {
     if (onSearch != null) {
       onSearch!(text).then((values) {
-        searchResults.clear();
-        searchResults.addAll(values);
+        _searchResults.clear();
+        _searchResults.addAll(values);
         notifyListeners();
       });
     } else {
-      searchResults.clear();
+      _searchResults.clear();
 
-      searchResults.addAll(
+      _searchResults.addAll(
         initData.where(
           (el) => el.text.toLowerCase().startsWith(
                 text.toLowerCase(),
@@ -101,6 +116,7 @@ class SearchableDropdownProvider<T> extends BaseDropdownProvider<T> {
 
   List<DropDownModel<T>> get getDropdownData {
     if (searchTextController.text.isNotEmpty || searchResults.isNotEmpty) {
+      _searchResults.removeWhere((el) => el == selectedValue);
       return searchResults;
     }
     return initData;