+++ lib/widgets/dropdown_container.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '/models/dropdown_decoration.dart';
+import '/df_dropdown.dart';
 import '../providers/base_dropdown_provider.dart';
 
 class DropdownContainer<T extends BaseDropdownProvider>
@@ -9,6 +9,7 @@ class DropdownContainer<T extends BaseDropdownProvider>
   const DropdownContainer({
     super.key,
     required this.disabled,
+    this.addDropdownKey = true,
     this.onTapInside,
     this.onTapOutside,
     this.labelText,
@@ -22,6 +23,7 @@ class DropdownContainer<T extends BaseDropdownProvider>
       vertical: 6,
       horizontal: 12,
     ),
+    required this.dropdownType,
   });
 
   final VoidCallback? onTapInside;
@@ -31,14 +33,17 @@ class DropdownContainer<T extends BaseDropdownProvider>
   final bool disableInput;
   final Widget? suffixWidget;
   final bool suffixTapEnabled;
+  final bool addDropdownKey;
   final Widget? child;
   final EdgeInsetsGeometry? contentPadding;
 
   final DropdownDecoration? decoration;
   final bool disabled;
+  final DropdownType dropdownType;
 
   @override
   Widget build(BuildContext context) {
+    final textTheme = Theme.of(context).textTheme;
     return IgnorePointer(
       ignoring: disabled,
       child: Opacity(
@@ -52,6 +57,7 @@ class DropdownContainer<T extends BaseDropdownProvider>
                 validator: provider.onValidateField,
               ),
               Container(
+                key: addDropdownKey ? provider.dropdownKey : null,
                 padding: contentPadding,
                 decoration: BoxDecoration(
                   color: decoration?.backgroundColor ?? Colors.transparent,
@@ -87,7 +93,10 @@ class DropdownContainer<T extends BaseDropdownProvider>
                     Expanded(
                       child: TapRegion(
                         onTapOutside: (_) {
-                          if (onTapOutside != null) onTapOutside!();
+                          if (onTapOutside != null &&
+                              provider.textFieldFocusNode.hasFocus) {
+                            onTapOutside!();
+                          }
                         },
                         onTapInside: (_) {
                           if (onTapInside != null) onTapInside!();
@@ -121,7 +130,25 @@ class DropdownContainer<T extends BaseDropdownProvider>
                       ),
                   ],
                 ),
-              ),
+              ), //An error message [Text] widget displayed only when validation returns an error
+
+              if (((dropdownType == DropdownType.expandable &&
+                          !provider.suggestionsExpanded) ||
+                      dropdownType == DropdownType.overlay) &&
+                  (decoration?.reserveSpaceForValidationMessage != false ||
+                      provider.validationError != null))
+                Padding(
+                  padding: const EdgeInsets.only(
+                    top: 4,
+                  ),
+                  child: Text(
+                    provider.validationError ?? "",
+                    style: decoration?.errorMessageTextStyle ??
+                        textTheme.bodySmall?.copyWith(
+                          color: Colors.red.shade500,
+                        ),
+                  ),
+                )
             ],
           ),
           child: child,